use simple setup for mysql endpoints (#885)
